This time the README file is fixed for sure! ^o^
trying to get a hang of this markdown syntax well it's better than
learning latex that's for sure. ^_^

made a simple script for windows

This is a simple windows perl script that is designed to move a file to
it's destination.

renamed to singleFileMover for obvious reasons

included latest script description!